+++ src/main/java/com/drmangotea/tfmg/base/events/TFMGCommonEvents.java
@@ -135,28 +135,4 @@ public class TFMGCommonEvents {
             }
         }
     }
-
-    @SubscribeEvent
-    public static void fixChemicaCylinders(ModifyDefaultComponentsEvent event) {
-        for (Item item : event.getAllItems().toList()) {
-            var key = BuiltInRegistries.ITEM.getKey(item);
-            if (key.getNamespace().equals("chemica")) {
-                CylinderFuels component = switch (key.getPath()) {
-                    case "biodiesel_engine_cylinder" -> new CylinderFuels(List.of(TFMGEngineFuelTypes.BIODIESEL));
-                    case "ethanol_engine_cylinder" -> new CylinderFuels(List.of(TFMGEngineFuelTypes.ETHANOL));
-                    case "high_cetane_engine_cylinder" -> new CylinderFuels(List.of(TFMGEngineFuelTypes.HIGH_CETANE_DIESEL));
-                    case "high_octane_engine_cylinder" -> new CylinderFuels(List.of(TFMGEngineFuelTypes.HIGH_OCTANE_GASOLINE));
-                    case "hydrogen_turbine_blade" -> new CylinderFuels(List.of(TFMGEngineFuelTypes.HYDROGEN_FUEL));
-                    default -> CylinderFuels.EMPTY;
-                };
-                if (!component.isEmpty()) {
-                    event.modify(item, (c) -> c.set(TFMGDataComponents.ENGINE_CYLINDER, component));
-                }
-                if (key.getPath().equals("platinum_electrode")) {
-                    Optional<Holder.Reference<Electrode>> electrodeHolder = TFMGRegistries.ELECTRODE_REGISTRY.getHolder(TFMG.asResource("chemica:electrode"));
-                    electrodeHolder.ifPresent(holder -> event.modify(item, (c) -> c.set(TFMGDataComponents.ELECTRODE, new Electrode.Stored(holder))));
-                }
-            }
-        }
-    }
 }