+++ src/main/java/com/drmangotea/tfmg/mixin/ItemStackMixin.java
@@ -1,0 +1,20 @@
+package com.drmangotea.tfmg.mixin;
+
+import com.drmangotea.tfmg.remap.TFMGRemapper;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ItemStack.class)
+public class ItemStackMixin {
+
+    @Inject(method = "inventoryTick", at = @At("HEAD"))
+    public void inventoryTick(Level level, Entity entity, int inventorySlot, boolean isCurrentItem, CallbackInfo ci) {
+        ItemStack thisStack = (ItemStack) (Object) this;
+        TFMGRemapper.remapComponents(thisStack, level.registryAccess());
+    }
+}